fixes very old bug in ModifyPredictor_WaterSpurtCap
This may result in signficant performance improvements, particularly
in big runs?  Worth comparing!

Also, this also highlights a poor performance decision made during
NLFV code introduction in all of the mpc_delagate_water functions.
GetDomainFaceValue() and SetDomainFaceValue() are not a good idea, as
they put GetComponent() calls into the innermost loop.  This should be
refactored with a lambda or function pointer or just plain duplicated
code to avoid doing this.